requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==3.7.1
blis==0.7.11
boto3==1.34.34
botocore==1.34.162
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cryptography==45.0.2
cymem==2.0.11
distro==1.9.0
ecdsa==0.19.1
environs==14.1.1
faiss-cpu==1.11.0
fastapi==0.104.1
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.2.2
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
jwt==1.3.1
langcodes==3.5.0
language_data==1.3.0
mangum==0.17.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
marshmallow==4.0.0
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pillow==11.2.1
preshed==3.0.9
pyasn1==0.4.8
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.7.1
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.27.0
sympy==1.14.0
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.24.0
wasabi==1.1.3
weasel==0.3.4
